% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal.forest.R
\name{causal.forest}
\alias{causal.forest}
\title{Causal forest}
\usage{
causal.forest(X, Y, W, sample.fraction = 0.5, mtry = ceiling(ncol(X)/3),
  num.trees = 500, num.threads = NULL, min.node.size = NULL,
  keep.inbag = FALSE, honesty = TRUE, ci.group.size = 4,
  precompute.nuisance = TRUE, seed = NULL)
}
\arguments{
\item{X}{The covariates used in the causal regression.}

\item{Y}{The outcome.}

\item{W}{The treatment assignment (may be binary or real).}

\item{sample.fraction}{Fraction of the data used to build each tree.
Note: If honesty is used, these subsamples will
further be cut in half.}

\item{mtry}{Number of variables tried for each split.}

\item{num.trees}{Number of trees grown in the forest. Note: Getting accurate
confidence intervals generally requires more trees than
getting accurate predictions.}

\item{num.threads}{Number of threads used in training. If set to NULL, the software
automatically selects an appropriate amount.}

\item{min.node.size}{Minimum number of observations in each tree leaf.}

\item{keep.inbag}{Currently not used.}

\item{honesty}{Should honest splitting (i.e., sub-sample splitting) be used?}

\item{ci.group.size}{The forst will grow ci.group.size trees on each subsample.
In order to provide confidence intervals, ci.group.size must
be at least 2.}

\item{precompute.nuisance}{Should we first run regression forests to estimate
y(x) = E[Y|X=x] and w(x) = E[W|X=x], and then run a
causal forest on the residuals? This approach is
recommended, computational resources permitting.}

\item{seed}{The seed of the c++ random number generator.}
}
\value{
A trained causal forest object.
}
\description{
Trains a causal forest that can be used to estimate
conditional average treatment effects tau(X). When
the treatment assignmnet W is binary and unconfounded,
we have tau(X) = E[Y(1) - Y(0) | X = x], where Y(0) and
Y(1) are potential outcomes corresponding to the two possible
treatment states. When W is continuous, we effectively estimate
an average partical effect Cov[Y, W | X = x] / Var[W | X = x],
and interpret it as a treatment effect given unconfoundedness.
}
