% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instrumental.forest.R
\name{instrumental.forest}
\alias{instrumental.forest}
\title{Intrumental forest}
\usage{
instrumental.forest(X, Y, W, Z, sample.fraction = 0.5,
  mtry = ceiling(ncol(X)/3), num.trees = 2000, num.threads = NULL,
  min.node.size = NULL, keep.inbag = FALSE, honesty = TRUE,
  ci.group.size = 2, precompute.nuisance = TRUE, split.regularization = 0,
  seed = NULL)
}
\arguments{
\item{X}{The covariates used in the instrumental regression.}

\item{Y}{The outcome.}

\item{W}{The treatment assignment (may be binary or real).}

\item{Z}{The instrument (may be binary or real).}

\item{sample.fraction}{Fraction of the data used to build each tree.
Note: If honesty is used, these subsamples will
further be cut in half.}

\item{mtry}{Number of variables tried for each split.}

\item{num.trees}{Number of trees grown in the forest. Note: Getting accurate
confidence intervals generally requires more trees than
getting accurate predictions.}

\item{num.threads}{Number of threads used in training. If set to NULL, the software
automatically selects an appropriate amount.}

\item{min.node.size}{Minimum number of observations in each tree leaf.}

\item{keep.inbag}{Currently not used.}

\item{honesty}{Should honest splitting (i.e., sub-sample splitting) be used?}

\item{ci.group.size}{The forst will grow ci.group.size trees on each subsample.
In order to provide confidence intervals, ci.group.size must
be at least 2.}

\item{precompute.nuisance}{Should we first run regression forests to estimate
y(x) = E[Y|X=x], w(x) = E[W|X=x] and z(x) = E[Z|X=x],
and then run an instrumental forest on the residuals?
This approach is recommended, computational resources
permitting.}

\item{split.regularization}{Whether splits should be regularized towards a naive
splitting criterion that ignores the instrument (and
instead emulates a causal forest).}

\item{seed}{The seed of the c++ random number generator.}
}
\value{
A trained instrumental forest object.
}
\description{
Trains an instrumental forest that can be used to estimate
conditional local average treatment effects tau(X) identified
using instruments. Formally, the forest estimates
tau(X) = Cov[Y, Z | X = x] / Cov[W, Z | X = x].
Note that when the instrument Z and treatment assignment W
coincide, an instrumental forest is equivalent to a causal forest.
}

