% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.forest.R
\name{regression.forest}
\alias{regression.forest}
\title{Regression forest}
\usage{
regression.forest(X, Y, sample.fraction = 0.5, mtry = ceiling(ncol(X)/3),
  num.trees = 500, num.threads = NULL, min.node.size = NULL,
  keep.inbag = FALSE, honesty = TRUE, ci.group.size = 2, seed = NULL)
}
\arguments{
\item{X}{The covariates used in the regression.}

\item{Y}{The outcome.}

\item{sample.fraction}{Fraction of the data used to build each tree.
Note: If honesty is used, these subsamples will
further be cut in half.}

\item{mtry}{Number of variables tried for each split.}

\item{num.trees}{Number of trees grown in the forest. Note: Getting accurate
confidence intervals generally requires more trees than
getting accurate predictions.}

\item{num.threads}{Number of threads used in training. If set to NULL, the software
automatically selects an appropriate amount.}

\item{min.node.size}{Minimum number of observations in each tree leaf.}

\item{keep.inbag}{Currently not used.}

\item{honesty}{Should honest splitting (i.e., sub-sample splitting) be used?}

\item{ci.group.size}{The forst will grow ci.group.size trees on each subsample.
In order to provide confidence intervals, ci.group.size must
be at least 2.}

\item{seed}{The seed of the c++ random number generator.}
}
\value{
A trained regression forest object.
}
\description{
Trains a regression forest that can be used to estimate
the conditional mean function mu(x) = E[Y | X = x]
}
